Start refactoring of NodesTable -> AddressTable
